Move state space creating into ModelBasedPlanningContext
